+++ apps/main/src/components/header/TaskDetailPageHeader.vue
@@ -39,6 +39,7 @@ const emit = defineEmits<{
   'add-to-album': [];
   help: [];
   'quick-settings': [];
+  'view-task-log': [];
   back: [];
 }>();
 
@@ -47,7 +48,15 @@ const showIds = computed(() => {
   if (IS_ANDROID) {
     return [HeaderFeatureId.Refresh, HeaderFeatureId.TaskDrawer];
   } else {
-    const ids = [HeaderFeatureId.Refresh, HeaderFeatureId.DeleteTask, HeaderFeatureId.AddToAlbum, HeaderFeatureId.TaskDrawer, HeaderFeatureId.Help, HeaderFeatureId.QuickSettings];
+    const ids = [
+      HeaderFeatureId.Refresh,
+      HeaderFeatureId.DeleteTask,
+      HeaderFeatureId.AddToAlbum,
+      HeaderFeatureId.TaskDrawer,
+      HeaderFeatureId.TaskViewLog,
+      HeaderFeatureId.Help,
+      HeaderFeatureId.QuickSettings,
+    ];
     if (props.showStopTask) ids.splice(1, 0, HeaderFeatureId.StopTask);
     return ids;
   }
@@ -55,7 +64,13 @@ const showIds = computed(() => {
 
 const foldIds = computed(() => {
   if (!IS_ANDROID) return [];
-  const ids = [HeaderFeatureId.DeleteTask, HeaderFeatureId.AddToAlbum, HeaderFeatureId.Help, HeaderFeatureId.QuickSettings];
+  const ids = [
+    HeaderFeatureId.DeleteTask,
+    HeaderFeatureId.AddToAlbum,
+    HeaderFeatureId.TaskViewLog,
+    HeaderFeatureId.Help,
+    HeaderFeatureId.QuickSettings,
+  ];
   if (props.showStopTask) ids.unshift(HeaderFeatureId.StopTask);
   return ids;
 });
@@ -80,6 +95,9 @@ const handleAction = (payload: { id: string; data: { type: string } }) => {
       break;
     case HeaderFeatureId.QuickSettings:
       emit("quick-settings");
+      break;
+    case HeaderFeatureId.TaskViewLog:
+      emit("view-task-log");
       break;
   }
 };